feat:insights project module init.

+++ lighthouse-insights/src/main/java/com/dtstep/lighthouse/insights/service/impl/ResourceServiceImpl.java
@@ -1,11 +1,13 @@
 package com.dtstep.lighthouse.insights.service.impl;
 
+import com.dtstep.lighthouse.common.exception.RoleDefendException;
 import com.dtstep.lighthouse.insights.dto.RolePair;
 import com.dtstep.lighthouse.insights.enums.ResourceTypeEnum;
 import com.dtstep.lighthouse.insights.enums.RoleTypeEnum;
 import com.dtstep.lighthouse.insights.modal.Resource;
 import com.dtstep.lighthouse.insights.modal.Role;
 import com.dtstep.lighthouse.insights.service.*;
+import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang3.Validate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -13,6 +15,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class ResourceServiceImpl implements ResourceService {
@@ -149,52 +152,32 @@ public class ResourceServiceImpl implements ResourceService {
         Role accessRole = null;
         if(resource.getResourceType() == ResourceTypeEnum.Department){
             manageRole = roleService.queryRole(RoleTypeEnum.DEPARTMENT_MANAGE_PERMISSION,resource.getResourceId());
-            if(roleService.isChildRoleExist(manageRole.getId())){
-                throw new Exception();
-            }
             accessRole = roleService.queryRole(RoleTypeEnum.DEPARTMENT_ACCESS_PERMISSION,resource.getResourceId());
-            if(roleService.isChildRoleExist(accessRole.getId())){
-                throw new Exception();
-            }
         }else if(resource.getResourceType() == ResourceTypeEnum.Project){
             manageRole = roleService.queryRole(RoleTypeEnum.PROJECT_MANAGE_PERMISSION,resource.getResourceId());
-            if(roleService.isChildRoleExist(manageRole.getId())){
-                throw new Exception();
-            }
             accessRole = roleService.queryRole(RoleTypeEnum.PROJECT_ACCESS_PERMISSION,resource.getResourceId());
-            if(roleService.isChildRoleExist(accessRole.getId())){
-                throw new Exception();
-            }
         }else if(resource.getResourceType() == ResourceTypeEnum.Group){
             manageRole = roleService.queryRole(RoleTypeEnum.GROUP_MANAGE_PERMISSION,resource.getResourceId());
-            if(roleService.isChildRoleExist(manageRole.getId())){
-                throw new Exception();
-            }
             accessRole = roleService.queryRole(RoleTypeEnum.GROUP_ACCESS_PERMISSION,resource.getResourceId());
-            if(roleService.isChildRoleExist(accessRole.getId())){
-                throw new Exception();
-            }
         }else if(resource.getResourceType() == ResourceTypeEnum.Stat){
             manageRole = roleService.queryRole(RoleTypeEnum.STAT_MANAGE_PERMISSION,resource.getResourceId());
-            if(roleService.isChildRoleExist(manageRole.getId())){
-                throw new Exception();
-            }
             accessRole = roleService.queryRole(RoleTypeEnum.STAT_ACCESS_PERMISSION,resource.getResourceId());
-            if(roleService.isChildRoleExist(accessRole.getId())){
-                throw new Exception();
-            }
         }else if(resource.getResourceType() == ResourceTypeEnum.Metric){
             manageRole = roleService.queryRole(RoleTypeEnum.METRIC_MANAGE_PERMISSION,resource.getResourceId());
-            if(roleService.isChildRoleExist(manageRole.getId())){
-                throw new Exception();
-            }
             accessRole = roleService.queryRole(RoleTypeEnum.METRIC_ACCESS_PERMISSION,resource.getResourceId());
-            if(roleService.isChildRoleExist(accessRole.getId())){
-                throw new Exception();
-            }
         }
         Validate.notNull(manageRole);
         Validate.notNull(accessRole);
+        List<Role> childManageRoles = roleService.queryListByPid(manageRole.getId(),1,20);
+        if(CollectionUtils.isNotEmpty(childManageRoles)){
+            String childRoles = childManageRoles.stream().map(z -> String.valueOf(z.getId())).collect(Collectors.joining(","));
+            throw new RoleDefendException("can't delete manage role [id:" + resource.getResourceId() + ",type:"+resource.getResourceType().name()+"],child roles ["+childRoles+"] exists!" );
+        }
+        List<Role> childAccessRoles = roleService.queryListByPid(manageRole.getId(),1,20);
+        if(CollectionUtils.isNotEmpty(childAccessRoles)){
+            String childRoles = childAccessRoles.stream().map(z -> String.valueOf(z.getId())).collect(Collectors.joining(","));
+            throw new RoleDefendException("can't delete access role [id:" + resource.getResourceId() + ",type:"+resource.getResourceType().name()+"],child roles ["+childRoles+"] exists!" );
+        }
         roleService.deleteById(manageRole.getId());
         roleService.deleteById(accessRole.getId());
     }
